hotfix: added s3 upload to create handler
