+++ src/main/java/no/unit/bibs/contents/CreateContentsApiHandler.java
@@ -12,14 +12,18 @@ import nva.commons.utils.StringUtils;
 import org.apache.http.HttpStatus;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
+
 import static java.util.Objects.isNull;
 
 public class CreateContentsApiHandler extends ApiGatewayHandler<ContentsRequest, GatewayResponse> {
 
     public static final String NO_PARAMETERS_GIVEN_TO_HANDLER = "No parameters given to CreateContentsApiHandler";
     public static final String COULD_NOT_INDEX_RECORD_PROVIDED = "Could not persist provided contents. ";
+    public static final String ERROR_STORING_FILE = "error storing file: ";
 
     private final DynamoDBClient dynamoDBClient;
+    private final S3Client s3Client;
 
     @JacocoGenerated
     public CreateContentsApiHandler() {
@@ -28,12 +32,13 @@ public class CreateContentsApiHandler extends ApiGatewayHandler<ContentsRequest,
 
     @JacocoGenerated
     public CreateContentsApiHandler(Environment environment) {
-        this(environment, new DynamoDBClient(environment));
+        this(environment, new DynamoDBClient(environment), new S3Client(environment));
     }
 
-    public CreateContentsApiHandler(Environment environment, DynamoDBClient dynamoDBClient) {
+    public CreateContentsApiHandler(Environment environment, DynamoDBClient dynamoDBClient, S3Client s3Client) {
         super(ContentsRequest.class, environment, LoggerFactory.getLogger(CreateContentsApiHandler.class));
         this.dynamoDBClient = dynamoDBClient;
+        this.s3Client = s3Client;
     }
 
 
@@ -41,7 +46,7 @@ public class CreateContentsApiHandler extends ApiGatewayHandler<ContentsRequest,
      * Implements the main logic of the handler. Any exception thrown by this method will be handled by {@link
      * RestRequestHandler#handleExpectedException} method.
      *
-     * @param request       The input object to the method. Usually a deserialized json.
+     * @param request     The input object to the method. Usually a deserialized json.
      * @param requestInfo Request headers and path.
      * @param context     the ApiGateway context.
      * @return the Response body that is going to be serialized in json
@@ -58,11 +63,21 @@ public class CreateContentsApiHandler extends ApiGatewayHandler<ContentsRequest,
         logger.error("json input looks like that :" + contentsDocument.toString());
         GatewayResponse gatewayResponse = new GatewayResponse(environment);
         if (StringUtils.isNotEmpty(contentsDocument.getIsbn())) {
-            logger.error("This is my IndexDocument to index: " + contentsDocument.toString());
-            dynamoDBClient.createContents(contentsDocument);
-            String createContents = dynamoDBClient.getContents(contentsDocument.getIsbn());
-            gatewayResponse.setBody(createContents);
-            gatewayResponse.setStatusCode(HttpStatus.SC_CREATED);
+
+            try {
+                s3Client.handleFiles(contentsDocument);
+                logger.error("This is my IndexDocument to index: " + contentsDocument.toString());
+                dynamoDBClient.createContents(contentsDocument);
+                String createContents = dynamoDBClient.getContents(contentsDocument.getIsbn());
+                gatewayResponse.setBody(createContents);
+                gatewayResponse.setStatusCode(HttpStatus.SC_CREATED);
+            } catch (IOException e) {
+                String msg = ERROR_STORING_FILE + e.getMessage();
+                logger.error(msg, e);
+                gatewayResponse.setErrorBody(msg);
+                gatewayResponse.setStatusCode(HttpStatus.SC_BAD_REQUEST);
+            }
+
         } else {
             logger.error(COULD_NOT_INDEX_RECORD_PROVIDED + contentsDocument.toString());
             gatewayResponse.setErrorBody(COULD_NOT_INDEX_RECORD_PROVIDED + contentsDocument.toString());
