+++ src/main/java/no/unit/bibs/contents/S3Client.java
@@ -6,6 +6,7 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.GeneratePresignedUrlRequest;
 import nva.commons.utils.Environment;
 import nva.commons.utils.JacocoGenerated;
+import nva.commons.utils.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,6 +15,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Date;
 
@@ -26,6 +28,19 @@ public class S3Client {
     public static final String CANNOT_CONNECT_TO_S3 = "Cannot connect to S3";
 
     public static final String CONTENT_DISPOSITION_FILENAME_TEMPLATE = "filename=\"%s\"";
+
+    public static final String ERROR_DOWNLOADING_FILE = "ISBN '%s' has invalid URL '%s' for file '%s' (%s): %s";
+    public static final String OBJECT_KEY_TEMPLATE = "/%s/%s/%s/%s";
+    public static final String FILE_NAME_TEMPLATE = "%s.%s";
+    public static final String FILE_EXTENSION_JPG = "jpg";
+    public static final String FILE_EXTENSION_WAV = "wav";
+    public static final String SMALL = "small";
+    public static final String LARGE = "large";
+    public static final String ORIGINAL = "original";
+    public static final String AUDIO = "audio";
+    public static final String MIME_TYPE_IMAGE_JPG = "image/jpg";
+    public static final String MIME_TYPE_AUDIO_WAV = "audio/wav";
+    public static final String HTTP_PREFIX = "http";
 
     private static final String AWS_REGION = "AWS_REGION";
     private static final String BUCKET_NAME = "BUCKET_NAME";
@@ -67,6 +82,91 @@ public class S3Client {
     }
 
     /**
+     * Uploads files found in ContentsDocument to S3 replacing url with s3 object key.
+     *
+     * @param contentsDocument contentsDocument
+     * @throws IOException IOException
+     */
+    @JacocoGenerated
+    public void handleFiles(ContentsDocument contentsDocument) throws IOException {
+        if (StringUtils.isNotEmpty(contentsDocument.getImageSmall())
+                && contentsDocument.getImageSmall().startsWith(HTTP_PREFIX)) {
+            String objectKey = putFileS3(
+                    contentsDocument.getIsbn(),
+                    contentsDocument.getImageSmall(),
+                    SMALL,
+                    FILE_EXTENSION_JPG,
+                    MIME_TYPE_IMAGE_JPG
+            );
+            contentsDocument.setImageSmall(objectKey);
+        }
+
+        if (StringUtils.isNotEmpty(contentsDocument.getImageLarge())
+                && contentsDocument.getImageLarge().startsWith(HTTP_PREFIX)) {
+            String objectKey = putFileS3(
+                    contentsDocument.getIsbn(),
+                    contentsDocument.getImageLarge(),
+                    LARGE,
+                    FILE_EXTENSION_JPG,
+                    MIME_TYPE_IMAGE_JPG
+            );
+            contentsDocument.setImageLarge(objectKey);
+        }
+
+        if (StringUtils.isNotEmpty(contentsDocument.getImageOriginal())
+                && contentsDocument.getImageOriginal().startsWith(HTTP_PREFIX)) {
+            String objectKey = putFileS3(
+                    contentsDocument.getIsbn(),
+                    contentsDocument.getImageOriginal(),
+                    ORIGINAL,
+                    FILE_EXTENSION_JPG,
+                    MIME_TYPE_IMAGE_JPG
+            );
+            contentsDocument.setImageOriginal(objectKey);
+        }
+
+        if (StringUtils.isNotEmpty(contentsDocument.getAudioFile())
+                && contentsDocument.getAudioFile().startsWith(HTTP_PREFIX)) {
+            String objectKey = putFileS3(
+                    contentsDocument.getIsbn(),
+                    contentsDocument.getAudioFile(),
+                    AUDIO,
+                    FILE_EXTENSION_WAV,
+                    MIME_TYPE_AUDIO_WAV
+            );
+            contentsDocument.setAudioFile(objectKey);
+        }
+    }
+
+    @JacocoGenerated
+    private String putFileS3(String isbn, String url, String type, String fileExtension, String mimeType)
+            throws IOException {
+
+        String fileName = String.format(FILE_NAME_TEMPLATE, isbn, fileExtension);
+        URL downloadUrl;
+        try {
+            downloadUrl = new URL(url);
+        } catch (MalformedURLException e) {
+            logger.error(String.format(ERROR_DOWNLOADING_FILE, isbn, url, fileName, type, e.getMessage()));
+            throw e;
+        }
+
+        String secondLinkPart = isbn.substring(isbn.length() - 2, isbn.length() - 1);
+        String firstLinkPart = isbn.substring(isbn.length() - 1);
+
+        String objectKey = String.format(OBJECT_KEY_TEMPLATE, type, firstLinkPart, secondLinkPart, fileName);
+        try (InputStream inputStream = downloadUrl.openStream()) {
+            uploadFile(
+                    inputStream,
+                    objectKey,
+                    fileName,
+                    mimeType
+            );
+        }
+        return objectKey;
+    }
+
+    /**
      * Uploads inputstream to S3 using a presigned upload write url.
      *
      * @param inputStream inputStream
@@ -77,7 +177,7 @@ public class S3Client {
      */
     @JacocoGenerated
     @SuppressWarnings("PMD.AssignmentInOperand")
-    public void uploadFile(InputStream inputStream, String objectName, String filename, String mimeType)
+    private void uploadFile(InputStream inputStream, String objectName, String filename, String mimeType)
             throws IOException {
         try {
             URL url = generatePresignedWriteUrl(objectName, filename, mimeType);
